+++ apps/toolbox/src/pages/list-page-sticky.ts
@@ -1,0 +1,7 @@
+import { EventData, Page } from '@nativescript/core';
+import { ListPageModelSticky } from './list-page-model-sticky';
+
+export function navigatingTo(args: EventData) {
+	const page = <Page>args.object;
+	page.bindingContext = new ListPageModelSticky();
+}
