+++ apps/toolbox/src/pages/list-page-sticky.xml
@@ -1,0 +1,23 @@
+<Page xmlns="http://schemas.nativescript.org/tns.xsd" navigatingTo="navigatingTo" actionBarHidden="false">
+    <Page.actionBar>
+        <ActionBar>
+            <Label text="Components" class="header"/>
+        </ActionBar>
+    </Page.actionBar>
+
+    <GridLayout backgroundColor="#efefef">
+        <ListView class="list-group" items="{{ countries }}" itemTap="{{ componentsItemTap }} " separatorColor="#00000000" itemTemplateSelector="{{ selectItemTemplate }}" stickyHeader="true" sectioned="true" stickyHeaderTopPadding="false" stickyHeaderTemplate="<GridLayout><Label text='{{ title }}' fontSize='18' fontWeight='bold' color='#009bff' padding='8 0 8 12' borderBottomWidth='1' borderBottomColor='#ccc' borderTopWidth='1' borderTopColor='#ccc' backgroundColor='#fff' /></GridLayout>" stickyHeaderHeight="45" itemLoading="{{ itemLoading }}" showSearch="true" searchAutoHide="true" searchChange="{{ onSearchTextChange }}">
+            <ListView.itemTemplates>
+                <template key="main">
+                    <GridLayout columns="auto,auto,*" padding="14 12 14 12" margin="2 6 2 6" borderRadius="10" backgroundColor="#fff" boxShadow="0px 1px 2px rgba(0,0,0,0.2)">
+                        <Label color="black" text="{{ flag }}" marginLeft="4"></Label>
+                        <Label col="1" text="{{ name }}" marginLeft="6" color="black"></Label>
+                        <Label col="2" text="{{ code }}" marginLeft="4" color="#999"></Label>
+                    </GridLayout>
+                </template>
+            </ListView.itemTemplates>
+        </ListView>
+
+    </GridLayout>
+
+</Page>
